+++ src/types/common/apiEndpoints.types.ts
