+++ src/types/common/page.types.ts
