+++ src/types/common/svg.d.ts
